+++ src/main/java/pageObjects/CalendarComponent.java
@@ -1,7 +1,9 @@
-package gmail.anastasiacoder.PageObjects;
+package pageObjects;
 
+import com.codeborne.selenide.ElementsCollection;
 import com.codeborne.selenide.SelenideElement;
 
+import static com.codeborne.selenide.Condition.exactText;
 import static com.codeborne.selenide.Condition.text;
 import static com.codeborne.selenide.Selenide.$;
 import static com.codeborne.selenide.Selenide.$$;
@@ -12,12 +14,12 @@ public class CalendarComponent {
         monthCalendar = $(".react-datepicker__month-select"),
         yearCalendar =  $(".react-datepicker__year-select");
 
-    private final String dayInCalendar = ".react-datepicker__day";
+    private final ElementsCollection daysInCalendar = $$(".react-datepicker__day");
 
     public void setDate(String day, String month, String year) {
         dateInput.click();
         monthCalendar.selectOption(month);
         yearCalendar.selectOption(year);
-        $$(dayInCalendar).find(text(day)).click();
+        daysInCalendar.find(exactText(day)).click();
     }
 }
