+++ src/test/java/gmail/anastasiacoder/test/base/TestBase.java
@@ -1,14 +1,9 @@
-package gmail.anastasiacoder.test;
+package gmail.anastasiacoder.test.base;
 
 import com.codeborne.selenide.Configuration;
-
-import gmail.anastasiacoder.PageObjects.RegistrationPage;
-
 import org.junit.jupiter.api.BeforeAll;
 
 public class TestBase {
-
-    RegistrationPage registrationPage = new RegistrationPage();
 
     @BeforeAll
     static void beforeAll() {
